+++ lib/drag_and_drop_builder_parameters.dart
@@ -18,22 +18,26 @@ class DragAndDropBuilderParameters {
       DragAndDropListInterface receiverList) onListReordered;
   final Axis axis;
   final CrossAxisAlignment verticalAlignment;
-  final double draggingWidth;
+  final double listDraggingWidth;
   final bool dragOnLongPress;
   final int itemSizeAnimationDuration;
   final Widget itemGhost;
   final double itemGhostOpacity;
   final Widget itemDivider;
+  final double itemDraggingWidth;
+  final Decoration itemDecorationWhileDragging;
   final int listSizeAnimationDuration;
   final Widget listGhost;
   final double listGhostOpacity;
   final EdgeInsets listPadding;
   final Decoration listDecoration;
+  final Decoration listDecorationWhileDragging;
   final Decoration listInnerDecoration;
   final double listWidth;
   final double lastItemTargetHeight;
   final bool addLastItemTargetHeightToTop;
   final Widget dragHandle;
+  final bool dragHandleOnLeft;
 
   DragAndDropBuilderParameters({
     @required this.onPointerMove,
@@ -42,7 +46,7 @@ class DragAndDropBuilderParameters {
     @required this.onItemReordered,
     @required this.onItemDropOnLastTarget,
     @required this.onListReordered,
-    @required this.draggingWidth,
+    @required this.listDraggingWidth,
     this.dragOnLongPress = true,
     this.axis = Axis.vertical,
     this.verticalAlignment = CrossAxisAlignment.start,
@@ -50,15 +54,19 @@ class DragAndDropBuilderParameters {
     this.itemGhostOpacity = 0.3,
     this.itemGhost,
     this.itemDivider,
+    this.itemDraggingWidth,
+    this.itemDecorationWhileDragging,
     this.listSizeAnimationDuration = 150,
     this.listGhostOpacity = 0.3,
     this.listGhost,
     this.listPadding,
     this.listDecoration,
+    this.listDecorationWhileDragging,
     this.listInnerDecoration,
     this.listWidth = double.infinity,
     this.lastItemTargetHeight = 20,
     this.addLastItemTargetHeightToTop = false,
     this.dragHandle,
+    this.dragHandleOnLeft = false,
   });
 }